# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmationRequest": {
        "additionalProperties": false,
        "properties": {
          "corrected_fields": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Correction"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision": {
            "enum": [
              "CONFIRM",
              "DISPUTE",
              "CORRECT"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "decision",
          "version"
        ],
        "title": "ConfirmationRequest",
        "type": "object"
      },
...
      "GraphEdge": {
        "description": "Invoice edges point from payer to receiver (issuer); so do transfer edges.",
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "invoice_count": {
            "default": 0,
            "title": "Invoice Count",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "INVOICE",
              "SETTLEMENT",
              "FX_LEG"
            ],
            "title": "Kind",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "currency",
          "amount_minor",
          "kind"
        ],
        "title": "GraphEdge",
        "type": "object"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "MEMBER",
              "SELF",
              "COUNTERPARTY",
              "MESH",
              "FX",
              "CARRY"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "kind"
        ],
        "title": "GraphNode",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "due_date": {
            "title": "Due Date",
            "type": "string"
          },
          "invoice_number": {
            "title": "Invoice Number",
            "type": "string"
          },
          "issue_date": {
            "title": "Issue Date",
            "type": "string"
          },
          "issuer_tax_id": {
            "title": "Issuer Tax Id",
            "type": "string"
          },
          "outstanding": {
            "default": "",
            "title": "Outstanding",
            "type": "string"
          },
          "payer_tax_id": {
            "title": "Payer Tax Id",
            "type": "string"
          }
        },
        "required": [
          "invoice_number",
          "issuer_tax_id",
          "payer_tax_id",
          "currency",
          "amount",
          "issue_date",
          "due_date"
        ],
        "title": "InvoiceCreate",
        "type": "object"
      },
      "InvoiceDetail": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "audit": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Audit",
            "type": "array"
          },
          "confirmations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Confirmations",
            "type": "array"
          },
          "counterparty": {
            "title": "Counterparty",
            "type": "string"
          },
          "counterparty_member_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Member Id"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "exclusions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Exclusions",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invoice_number": {
            "title": "Invoice Number",
            "type": "string"
          },
          "issue_date": {
            "format": "date",
            "title": "Issue Date",
            "type": "string"
          },
          "issuer_tax_id": {
            "title": "Issuer Tax Id",
            "type": "string"
          },
          "needs_confirmation": {
            "title": "Needs Confirmation",
            "type": "boolean"
          },
          "outcomes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Outcomes",
            "type": "array"
          },
          "outstanding_minor": {
            "title": "Outstanding Minor",
            "type": "integer"
          },
          "payer_tax_id": {
            "title": "Payer Tax Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "invoice_number",
          "issue_date",
          "due_date",
          "currency",
          "amount_minor",
          "outstanding_minor",
          "status",
          "current_version",
          "direction",
          "counterparty",
          "counterparty_member_id",
          "issuer_tax_id",
          "payer_tax_id",
          "needs_confirmation",
          "versions",
          "confirmations",
          "exclusions",
          "outcomes",
          "audit"
        ],
        "title": "InvoiceDetail",
        "type": "object"
      },
      "InvoiceView": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "counterparty": {
            "title": "Counterparty",
            "type": "string"
          },
          "counterparty_member_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Member Id"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "invoice_number": {
            "title": "Invoice Number",
            "type": "string"
          },
          "issue_date": {
            "format": "date",
            "title": "Issue Date",
            "type": "string"
          },
          "issuer_tax_id": {
            "title": "Issuer Tax Id",
            "type": "string"
          },
          "needs_confirmation": {
            "title": "Needs Confirmation",
            "type": "boolean"
          },
          "outstanding_minor": {
            "title": "Outstanding Minor",
            "type": "integer"
          },
          "payer_tax_id": {
            "title": "Payer Tax Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "invoice_number",
          "issue_date",
          "due_date",
          "currency",
          "amount_minor",
          "outstanding_minor",
          "status",
          "current_version",
          "direction",
          "counterparty",
          "counterparty_member_id",
          "issuer_tax_id",
          "payer_tax_id",
          "needs_confirmation"
        ],
        "title": "InvoiceView",
        "type": "object"
      },
...
      "NetworkView": {
        "properties": {
          "attempt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt"
          },
          "gross_minor": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Gross Minor",
            "type": "object"
          },
          "invoice_edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Invoice Edges",
            "type": "array"
          },
          "net_minor": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Net Minor",
            "type": "object"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Status"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunChoice"
            },
            "title": "Runs",
            "type": "array"
          },
          "source": {
            "enum": [
              "RUN",
              "WINDOW",
              "MEMBER"
            ],
            "title": "Source",
            "type": "string"
          },
          "transfer_edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Transfer Edges",
            "type": "array"
          }
        },
        "required": [
          "source",
          "run_id",
          "run_status",
          "attempt",
          "nodes",
          "invoice_edges",
          "transfer_edges",
          "gross_minor",
          "net_minor",
          "runs"
        ],
        "title": "NetworkView",
        "type": "object"
      },
      "NotificationList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationView"
            },
            "title": "Items",
            "type": "array"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "unread_count"
        ],
        "title": "NotificationList",
        "type": "object"
      },
...
      "Page_InvoiceView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[InvoiceView]",
        "type": "object"
      },
...
      "UploadResult": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/UploadError"
            },
            "title": "Errors",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceView"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "errors"
        ],
        "title": "UploadResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multilateral invoice netting and settlement (hackathon prototype).",
    "title": "Wise Mesh Continental API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/network": {
      "get": {
        "description": "The whole graph: gross invoice edges before netting, settlement transfers after.",
        "operationId": "admin_network_v1_admin_network_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Network",
        "tags": [
          "network"
        ]
      }
    },
...
    "/v1/invoices": {
      "get": {
        "operationId": "list_invoices_v1_invoices_get",
        "parameters": [
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "RECEIVABLE",
                    "PAYABLE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "counterparty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counterparty"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InvoiceView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Invoices",
        "tags": [
          "invoices"
        ]
      },
      "post": {
        "operationId": "create_v1_invoices_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/v1/invoices/uploads": {
      "post": {
        "operationId": "upload_v1_invoices_uploads_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/invoices/{invoice_id}": {
      "get": {
        "operationId": "detail_v1_invoices__invoice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/invoices/{invoice_id}/confirmations": {
      "post": {
        "operationId": "decide_v1_invoices__invoice_id__confirmations_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/v1/notifications": {
      "get": {
        "operationId": "notifications_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_read_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
  }
}